Add input options related to `copy()` (#161)
* Add absolute_build_dir input option

This commit adds absolute_build_dir as an additional input. If the
option is set, the provided build_dir argument won't be treated as a
relative path to the current working directory anymore. This is helpful
in environments like Nix, where the build is output to a static folder
(for example /nix/store).

Closes #159.

* Add follow_symlinks option

This option allows symbolic links in the source directory to be
followed, recursively copying the entire directory structure.

* Update outputs

* Add missing inputs to action.yml

* Add default values to input options

This concerns absolute_build_dir and follow_symlinks.

Co-authored-by: CrazyMax <github@crazymax.dev>

Co-authored-by: CrazyMax <github@crazymax.dev>